Write backend for task 2 and updates for 1
Write out code for task 2, finishing most of it. Pretty much the major things left to do are all on the front end side of things, such as building the pages and styling them to look not-awful. Might need a few more backend changes, but going to break again and come back to it later.